requirements.txt
djangorestframework~=3.14.0
Django==4.2.7
pika==1.3.2
prettytable==3.9.0
requests==2.31.0
